Convert 'drop' statement to only drop existing table
After building the codebase as described in the Readme, and running the pipeline tools as described (`./build/tkgetxml && ./build/tkconv && ./build/tkpull && ./build/tkindex`), the `tkconv` utility fails when attempting to delete a non-existing table.

> Render queries..
> terminate called after throwing an instance of 'std::runtime_error'
>   what():  Unable to prepare query drop table openvragen: no such table: openvragen
> Aborted

Most likely, a prior termination of one of the components left the database in an inconsistent state. Replacing the DROP statement on line 937 of tkconv.cc with

>   sqlw.query("drop table if exists openvragen");

fixed the issue for me.